[TEMPORARY] unlock view methods and fix in associator

+++ DataFormats/TICL/interface/AssociationMap.h
@@ -30,7 +30,7 @@ namespace ticl {
 							auto offset = (key == 0u) ? 0u : this->offsets()[key].keys_offsets();
 							auto size = (key == 0u) ? this->offsets()[0].keys_offsets() 
                                         : this->offsets()[key].keys_offsets() - this->offsets()[key - 1].keys_offsets();
-							return std::span<TMapped>{this->content().values().data() + offset, size};
+							return std::span<TMapped>{this->content().values().data() + offset, static_cast<std::size_t>(size)};
 							}
 						),
                         SOA_CONST_VIEW_METHODS(
@@ -38,7 +38,7 @@ namespace ticl {
 							auto offset = (key == 0) ? 0u : this->offsets().keys_offsets()[key];
 							auto size = (key == 0u) ? this->offsets()[0].keys_offsets()
                                         : this->offsets()[key].keys_offsets() - this->offsets()[key - 1].keys_offsets();
-							return std::span<const TMapped>{this->content().values().data() + offset, size};
+							return std::span<const TMapped>{this->content().values().data() + offset, static_cast<std::size_t>(size)};
 							}
 							constexpr SOA_HOST_DEVICE auto contains(TKey key) const {
 							return this->count(key) > 0;
